# gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Pack Optimizer
    description: Service to calculate optimal pack combinations for a given number of items
    version: 0.0.1
servers:
    - url: http://localhost:8080
paths:
    /health:
        get:
            tags:
                - optimizer
            summary: health optimizer
            description: Returns the health status of the service.
            operationId: optimizer#health
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HealthResponseBody'
                            example:
                                status: ok
                "500":
                    description: 'internal_server_error: Internal Server Error response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /packs/calculate:
        get:
            tags:
                - optimizer
            summary: calculate optimizer
            description: Calculate optimal pack combinations for a given number of items.
            operationId: optimizer#calculate
            parameters:
                - name: quantity
                  in: query
                  description: Total number of items to pack
                  allowEmptyValue: true
                  required: true
                  schema:
                    type: integer
                    description: Total number of items to pack
                    example: 500
                    format: int64
                    minimum: 1
                  example: 500
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CalculateResponseBody'
                            example:
                                packs:
                                    - quantity: 1609558464246327244
                                      size: 5614756764895872828
                                    - quantity: 1609558464246327244
                                      size: 5614756764895872828
                "400":
                    description: 'bad_request: Bad Request response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "500":
                    description: 'internal_server_error: Internal Server Error response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /packs/sizes:
        get:
            tags:
                - optimizer
            summary: getPackSizes optimizer
            description: Get the current pack sizes.
            operationId: optimizer#getPackSizes
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetPackSizesResponseBody'
                            example:
                                sizes:
                                    - 250
                                    - 500
                                    - 1000
                "500":
                    description: 'internal_server_error: Internal Server Error response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
        put:
            tags:
                - optimizer
            summary: updatePackSizes optimizer
            description: Update the pack sizes.
            operationId: optimizer#updatePackSizes
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GetPackSizesResponseBody'
                        example:
                            sizes:
                                - 250
                                - 500
                                - 1000
            responses:
                "204":
                    description: No Content response.
                "400":
                    description: 'bad_request: Bad Request response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "500":
                    description: 'internal_server_error: Internal Server Error response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
components:
    schemas:
        CalculateResponseBody:
            type: object
            properties:
                packs:
                    type: array
                    items:
                        $ref: '#/components/schemas/Pack'
                    description: Optimal pack combinations (pack size -> quantity)
                    example:
                        - quantity: 1609558464246327244
                          size: 5614756764895872828
                        - quantity: 1609558464246327244
                          size: 5614756764895872828
                        - quantity: 1609558464246327244
                          size: 5614756764895872828
                        - quantity: 1609558464246327244
                          size: 5614756764895872828
            example:
                packs:
                    - quantity: 1609558464246327244
                      size: 5614756764895872828
                    - quantity: 1609558464246327244
                      size: 5614756764895872828
                    - quantity: 1609558464246327244
                      size: 5614756764895872828
                    - quantity: 1609558464246327244
                      size: 5614756764895872828
            required:
                - packs
        Error:
            type: object
            properties:
                fault:
                    type: boolean
                    description: Is the error a server-side fault?
                    example: true
                id:
                    type: string
                    description: ID is a unique identifier for this particular occurrence of the problem.
                    example: 123abc
                message:
                    type: string
                    description: Message is a human-readable explanation specific to this occurrence of the problem.
                    example: parameter 'p' must be an integer
                name:
                    type: string
                    description: Name is the name of this class of errors.
                    example: bad_request
                temporary:
                    type: boolean
                    description: Is the error temporary?
                    example: false
                timeout:
                    type: boolean
                    description: Is the error a timeout?
                    example: true
            example:
                fault: true
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: false
                timeout: false
            required:
                - name
                - id
                - message
                - temporary
                - timeout
                - fault
        GetPackSizesResponseBody:
            type: object
            properties:
                sizes:
                    type: array
                    items:
                        type: integer
                        example: 3314533960745271680
                        format: int64
                    description: Current pack sizes
                    example:
                        - 250
                        - 500
                        - 1000
            example:
                sizes:
                    - 250
                    - 500
                    - 1000
            required:
                - sizes
        HealthResponseBody:
            type: object
            properties:
                status:
                    type: string
                    description: Health status of the service
                    example: ok
            example:
                status: ok
            required:
                - status
        Pack:
            type: object
            properties:
                quantity:
                    type: integer
                    description: Quantity of this pack size needed
                    example: 5508737646480335232
                    format: int64
                size:
                    type: integer
                    description: Size of the pack
                    example: 2315902810084718852
                    format: int64
            description: Pack represents a pack size and the quantity needed.
            example:
                quantity: 1983071130868209503
                size: 7969886467026274248
            required:
                - size
                - quantity
tags:
    - name: optimizer
      description: Service calculates optimal pack combinations for a given number of items.